msm: camera: isp: Fix warning and errors based on static analysis
This change fixes the warning/errors from static analysis

Conflicts:
        drivers/media/platform/msm/camera_v2/isp/msm_isp_axi_util.c
        drivers/media/platform/msm/camera_v2/isp/msm_isp_util.c
        drivers/media/platform/msm/camera_v2/ispif/msm_ispif.c

CRs-fixed: 992942

Conflicts:
	drivers/media/platform/msm/camera_v2/isp/msm_isp_axi_util.c
	drivers/media/platform/msm/camera_v2/isp/msm_isp_util.c
	drivers/media/platform/msm/camera_v2/ispif/msm_ispif.c

+++ drivers/media/platform/msm/camera_v2/ispif/msm_ispif.c
@@ -379,9 +379,9 @@ static void msm_ispif_enable_intf_cids(struct ispif_device *ispif,
 
 	data = msm_camera_io_r(ispif->base + intf_addr);
 	if (enable)
-		data |= cid_mask;
+		data |= (uint32_t)cid_mask;
 	else
-		data &= ~cid_mask;
+		data &= ~((uint32_t)cid_mask);
 	msm_camera_io_w_mb(data, ispif->base + intf_addr);
 }
